+++ Vault/src/utopia/vault/nosql/access/ManyIdModelAccess.scala
@@ -11,7 +11,7 @@ import utopia.vault.sql.Extensions._
  * @param ids Row ids to target
  * @param factory Factory used when reading model data
  */
-class ManyIdModelAccess[+A](ids: Set[Value], override val factory: FromResultFactory[A]) extends ManyModelAccess[A]
+class ManyIdModelAccess[+A](ids: Iterable[Value], override val factory: FromResultFactory[A]) extends ManyModelAccess[A]
 {
 	// COMPUTED --------------------------